refactor(database): move AppDatabase to com.akilimo.mobile.database package
Relocates AppDatabase from the root package into the database subpackage
alongside DatabaseMigrations, and updates all 8 import sites accordingly.
