+++ app/src/main/java/com/akilimo/mobile/base/BaseActivity.kt
@@ -18,7 +18,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.viewbinding.ViewBinding
 import com.akilimo.mobile.AkilimoApp
-import com.akilimo.mobile.AppDatabase
+import com.akilimo.mobile.database.AppDatabase
 import com.akilimo.mobile.data.AppSettingsDataStore
 import com.akilimo.mobile.data.AppSettingsEntryPoint
 import com.akilimo.mobile.helper.LocaleHelper
